pyproject.toml
[project]
name = "simple-image-storage"
version = "0.1.0"
description = "A simple api that can store and serve images written in python with fastapi"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "uvicorn-worker>=0.3.0",
]
